WIP - Bind in server message to client
The client now sends a message when a user clicks on a pill button to
decide which view to render of the final results. This message is split
into subtype at the server level, and an appropriate reply is sent.
Also added function to be able to show multiple sections at once when
their names are passed as arguments at client level.

TODO: finalize parsing of the annotated message at the client level to
render adequately.